fix: keep setup markdown.rs under the U-16 hard limit
Extract Claude settings hook helpers into a sibling module so Core CI
validate-u16-baseline stops failing on the 800-line hard limit.
